update new structure for access control (#65)
* update new structure for access control

* update test for new access control

* update userGuide for the new access control

* Integrate with shannon's code

* update test for access control

* update tests

* Integration with bangjie's code

* integrate with outletinformation to access password.

+++ src/main/java/seedu/ptman/logic/commands/DeleteCommand.java
@@ -1,6 +1,7 @@
 package seedu.ptman.logic.commands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.ptman.commons.core.Messages.MESSAGE_ACCESS_DENIED;
 
 import java.util.List;
 import java.util.Objects;
@@ -32,13 +33,17 @@ public class DeleteCommand extends UndoableCommand {
 
     public DeleteCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
-        isAdminCommand = true;
     }
 
 
     @Override
-    public CommandResult executeUndoableCommand() {
+    public CommandResult executeUndoableCommand() throws CommandException {
         requireNonNull(employeeToDelete);
+
+        if (!model.isAdminMode()) {
+            throw new CommandException(MESSAGE_ACCESS_DENIED);
+        }
+
         try {
             model.deleteEmployee(employeeToDelete);
         } catch (EmployeeNotFoundException pnfe) {
